Fixes an issue with `isIovLedgerAppSignature`
Inside the `sign()` method of the ledger app class there's a call to
`isIovLedgerAppSignature`. The provided object has no `signature`
attribute and since the `isIovLedgerAppSignature` method is evaluating
`signature.length` directly then there was an exception about trying to
use `length` on `undefined`.

This fix makes what I think is a more robust method to check if the
object is actually a signature or an error object, starting by checking
precisely that the `returnCode` is not an error code and then doing some
sanity check on the `signature` attribute